fashion-mnist-ml-system/requirements/frontend.txt
streamlit==1.37.1
requests==2.32.3
numpy==1.26.4
pandas==2.2.2
pillow==10.4.0
